chore: document Raffles
chore: document raffleModalSubmits
+ extract updateEntryTotal() that was duplicated in raffleButtonSubmits

chore: document raffleMessageContextSubmits

chore: document raffleButtonSubmits
